+++ devenv.nix
@@ -24,14 +24,4 @@
   scripts.generate-env-file.exec = ''
     terraform output -raw devenv_local_nix > devenv.local.nix
   '';
-
-  scripts.proxy-vps.exec = ''
-    exec sshuttle -e "ssh -o CheckHostIP=no -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null" -r "$(terraform output --raw vm_username)@$(terraform output --raw vm_ip)" 0/0
-  '';
-
-  scripts.start-app-service-ssh.exec = ''
-      resource_group=$(terraform output --raw "resource_group")
-      app_name=$(terraform output --raw "app_service_app_name")
-      exec az webapp create-remote-connection --resource-group $resource_group --name $app_name
-  '';
 }