+++ test/support/azure_vm_vpn.rb
@@ -4,8 +4,9 @@ require "shellwords"
 class AzureVmVpn
   def initialize(verbose: false)
     @verbose = verbose
-    stdin, stdout, @wait_thread = Open3.popen2e("proxy-vps")
+    stdin, stdout, @wait_thread = Open3.popen2e("bin/proxy-vps")
     stdout.each do |line|
+      puts line if @verbose
       break if line.include?("Connected to server")
     end
   end